Plot synteny between a query and target sequence in a paf alignment

Plot synteny between a query and target sequence in a paf alignment

plot_synteny(ali, q_chrom, t_chrom, centre = TRUE, rc = FALSE,
  xlab = "Position in query", ylab = "", x_labeller = Mb_lab)

Arguments

ali
pafr or tibble containing the genome alignment (as returned by read_paf)

q_chrom
character, name for the query sequence

t_chrom
character, name for the target sequence

centre,
logical, if TRUE (default) adjust the position of the target sequence so it is centred on the query. If not, both sequences start at position zero

rc
logical, if TRUE, use the reverse and complement for the target sequence

xlab,
character, Name for for x-axis

ylab,
character, Name for for x-axus

x_labeller
function to be used to label x-axis (defaults to

Value

A ggplot object displaying synteny between query and target sequences

Examples

ali <- read_paf( system.file("extdata", "fungi.paf", package = "pafr") )
plot_syntenty(q_chrom="Q_chr3", t_chrom="T_chr4", centre=TRUE)

#> Error in plot_syntenty(q_chrom = "Q_chr3", t_chrom = "T_chr4", centre = TRUE): could not find function "plot_syntenty"

plot_synteny(long_ali, q_chrom="Q_chr5", t_chrom="T_chr5", centre=TRUE)

#> Error in subset(ali, qname == q_chrom & tname == t_chrom): object 'long_ali' not found

plot_synteny(long_ali, q_chrom="Q_chr5", t_chrom="T_chr5", centre=TRUE, rc=TRUE)

#> Error in subset(ali, qname == q_chrom & tname == t_chrom): object 'long_ali' not found